move AndrolidExceptions to the package brut.androlib.exceptions (renamed from brut.androlib.err)
