+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/exceptions/OutDirExistsException.java
@@ -14,12 +14,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package brut.androlib.err;
+package brut.androlib.exceptions;
 
-import brut.androlib.AndrolibException;
-
-public class RawXmlEncounteredException extends AndrolibException {
-    public RawXmlEncounteredException(String message, Throwable cause) {
-        super(message, cause);
-    }
+public class OutDirExistsException extends AndrolibException {
+	public OutDirExistsException() {
+	}
 }
