+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/exceptions/UndefinedResObjectException.java
@@ -14,11 +14,10 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package brut.androlib.err;
+package brut.androlib.exceptions;
 
-import brut.androlib.AndrolibException;
-
-public class OutDirExistsException extends AndrolibException {
-	public OutDirExistsException() {
+public class UndefinedResObjectException extends AndrolibException {
+	public UndefinedResObjectException(String message) {
+		super(message);
 	}
 }
