+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/data/ResResSpec.java
@@ -16,8 +16,8 @@
  */
 package brut.androlib.res.data;
 
-import brut.androlib.AndrolibException;
-import brut.androlib.err.UndefinedResObjectException;
+import brut.androlib.exceptions.AndrolibException;
+import brut.androlib.exceptions.UndefinedResObjectException;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Arrays;
