+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/data/ResResource.java
@@ -16,7 +16,7 @@
  */
 package brut.androlib.res.data;
 
-import brut.androlib.AndrolibException;
+import brut.androlib.exceptions.AndrolibException;
 import brut.androlib.res.data.value.ResValue;
 
 public class ResResource {
