+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/data/value/ResDimenValue.java
@@ -17,7 +17,7 @@
 package brut.androlib.res.data.value;
 
 import android.util.TypedValue;
-import brut.androlib.AndrolibException;
+import brut.androlib.exceptions.AndrolibException;
 
 public class ResDimenValue extends ResIntValue {
     public ResDimenValue(int value, String rawValue) {
