+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/data/value/ResEmptyValue.java
@@ -16,7 +16,7 @@
  */
 package brut.androlib.res.data.value;
 
-import brut.androlib.AndrolibException;
+import brut.androlib.exceptions.AndrolibException;
 
 public class ResEmptyValue extends ResScalarValue {
     protected final int mValue;
