+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/data/value/ResFileValue.java
@@ -16,7 +16,7 @@
  */
 package brut.androlib.res.data.value;
 
-import brut.androlib.AndrolibException;
+import brut.androlib.exceptions.AndrolibException;
 
 public class ResFileValue extends ResIntBasedValue {
     private final String mPath;
