+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/xml/ResXmlEncodable.java
@@ -16,7 +16,7 @@
  */
 package brut.androlib.res.xml;
 
-import brut.androlib.AndrolibException;
+import brut.androlib.exceptions.AndrolibException;
 
 public interface ResXmlEncodable {
     String encodeAsResXmlAttr() throws AndrolibException;
