+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/xml/ResXmlPatcher.java
@@ -16,7 +16,7 @@
  */
 package brut.androlib.res.xml;
 
-import brut.androlib.AndrolibException;
+import brut.androlib.exceptions.AndrolibException;
 import org.w3c.dom.*;
 import org.xml.sax.SAXException;
